Move: LLMCompressorPlugin into it's own submodule
